app/requirements.txt
Django==4.2.4
django-celery-results==2.5.1.0
django-encrypted-model-fields==0.6.5
django-phonenumber-field==7.1.0
django-import-export==3.2.0
# keras==2.11.0
# keras-cv==0.3.4
# tensorflow==2.11.0
# tensorflow-datasets==4.8.2
selenium==4.11.2
selenium-wire==5.1.0
schwifty==2023.3.0
PyICU==2.11
python-telegram==0.16.0
phonenumberslite==8.13.19
celery==5.3.4
dateparser==1.1.8
psycopg2==2.9.7
Pillow==9.5.0
holehe==1.61
geopandas==0.12.2
geopy==2.4.0
Shapely==1.8.5.post1
cryptography==41.0.3
redis==4.6.0
pyee==9.1.0
aiohttp==3.8.5
fake-useragent==1.2.1
Faker==18.7.0
dataclasses-json==0.5.14
folium==0.14.0
matplotlib==3.7.2
mapclassify==2.6.0
django-tinymce==3.6.1